Adds JEMALLOC_ENABLED config option
This is a quality of life addition that makes it easier to configure
LD_PRELOAD. It appears the library name changed between 4.x and 5.x
which looks to have caused some confusion. It also seems that getting it
in the Heroku dashboard acts differently from setting it on the command
line (I think).