Fix error when JEMALLOC_ENABLED is not set
When loading profile.d/jemalloc we would get the below error. It boils
down to asking bash to do a string comparison when the left side wasn't
a string. The fix is to force the left side to be a string. So, if it's
unset, the comparison becomes `"" = "true"`.

```
/app/.profile.d/jemalloc.sh: line 5: [: =: unary operator expected
```

Fixes #1