Removes static libs and include files from the buildpack
This reduces the bundle from ~8.2MB for 5.1.0 to 1.13MB which will save
bandwidth and slug size. The biggest saving came from removing the
static libs which weighed in around 45MB uncompressed. I'm not aware of
a usecase where you would prefer the static builds over the shared, so
if you're reading this wondering "why were these removed" then feel free
to open an issue with your use case.

This also removes the include headers. They're really small and compress
well but they're only useful when compiling code against jemalloc. I've
taken step away from this buildpack being useful for compiling other
buildpacks against it, this continues that direction.

Lastly, this applies to the builds for 5.1.0 and won't be backported to
earlier builds unless there is another reason to rebuild them. I would
prefer builds to be stable once uploaded and depended on by folks.