Check the dyno's STACK to download the correct version on stack change
As edmorley points out, there can be ABI changes between different
stacks which could cause issues. It is, after all, the reason that we
have a download specific to each stack.

This change adds a .stack file in the cached directory which we use to
track which version of the compiled binary was originally downloaded.
When that file is missing (i.e. for anyone already using the buildpack)
then we will force a download on the next deploy.

Fixes #19